pyproject.toml
[project]
name = "tfah"
version = "0.1.0"
description = "Transient Fault Auto-Healer — Autonomous SRE agent for transient fault remediation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Agentic framework
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    # Mock server
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    # Vulnerable app / resilience
    "requests>=2.32.0",
    "tenacity>=9.0.0",
    # GitHub automation
    "PyGithub>=2.5.0",
    # Notifications
    "httpx>=0.28.0",
    # Testing
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    # Env
    "python-dotenv>=1.0.0",
]
